+++ packages/chain-adapters/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [@shapeshiftoss/chain-adapters-v10.7.1](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v10.7.0...@shapeshiftoss/chain-adapters-v10.7.1) (2023-01-05)
+
+
+### Bug Fixes
+
+* only inject accountNumber, not full bip44params ([#1147](https://github.com/shapeshift/lib/issues/1147)) ([7e53b6d](https://github.com/shapeshift/lib/commit/7e53b6d517d069d8a1831e4b474405b4e9220dcf))
+
+# [@shapeshiftoss/chain-adapters-v10.7.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v10.6.0...@shapeshiftoss/chain-adapters-v10.7.0) (2023-01-05)
+
+
+### Features
+
+* **chain-adapters:UtxoBaseAdapter:** allow UTXOs selection from `from` param ([#1148](https://github.com/shapeshift/lib/issues/1148)) ([ea9cede](https://github.com/shapeshift/lib/commit/ea9cedea8f99e684ffa00df46b634b62cd156450))
+
 # [@shapeshiftoss/chain-adapters-v10.6.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v10.5.1...@shapeshiftoss/chain-adapters-v10.6.0) (2022-12-23)