+++ packages/swapper/CHANGELOG.md
@@ -1,3 +1,24 @@
+# [@shapeshiftoss/swapper-v15.1.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/swapper-v15.0.4...@shapeshiftoss/swapper-v15.1.0) (2023-01-05)
+
+
+### Features
+
+* **swapper:** update ThornodePoolResponse to the latest /lcd/thorchain/pools schema ([#1152](https://github.com/shapeshift/lib/issues/1152)) ([500af15](https://github.com/shapeshift/lib/commit/500af150287f820b8719bd67c16b75e707e184b2))
+
+# [@shapeshiftoss/swapper-v15.0.4](https://github.com/shapeshift/lib/compare/@shapeshiftoss/swapper-v15.0.3...@shapeshiftoss/swapper-v15.0.4) (2023-01-05)
+
+
+### Bug Fixes
+
+* only inject accountNumber, not full bip44params ([#1147](https://github.com/shapeshift/lib/issues/1147)) ([7e53b6d](https://github.com/shapeshift/lib/commit/7e53b6d517d069d8a1831e4b474405b4e9220dcf))
+
+# [@shapeshiftoss/swapper-v15.0.3](https://github.com/shapeshift/lib/compare/@shapeshiftoss/swapper-v15.0.2...@shapeshiftoss/swapper-v15.0.3) (2023-01-05)
+
+
+### Performance Improvements
+
+* **swapper:** add caching to swapper requests ([#1151](https://github.com/shapeshift/lib/issues/1151)) ([76c6f84](https://github.com/shapeshift/lib/commit/76c6f84f496931aad3c58b8f4088f84287e426ed))
+
 # [@shapeshiftoss/swapper-v15.0.2](https://github.com/shapeshift/lib/compare/@shapeshiftoss/swapper-v15.0.1...@shapeshiftoss/swapper-v15.0.2) (2022-12-23)